Add create goal functionality to CLI
This commit adds the minimum to complete a goal entry from the CLI the
code works but can be refactored and built upon for instance it could
easily support having an entry for blocks feature.  With that case
a refactor would definitely be in order but I wanted something I could
use right away and I do find this tool very useful. Any feedback is
definitely appreciated.